+++ Xamarin.Forms.Platform.Android/Renderers/ShellFlyoutRenderer.cs
@@ -185,7 +185,7 @@ namespace Xamarin.Forms.Platform.Android
 
 			unchecked
 			{
-				SetScrimColor(behavior == FlyoutBehavior.Locked ? Color.Transparent.ToAndroid() : (int)DefaultScrimColor);
+				SetScrimColor(behavior == FlyoutBehavior.Locked ? (int)Color.Transparent.ToAndroid() : (int)DefaultScrimColor);
 			}
 		}
